requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
Pillow>=10.0.0
